Implement the WHATWG URL and URLSearchParam classes
This is a full implementation of the WHATWG URL and URLSearchParam classes, based on the [rust-url](https://crates.io/crates/url), which is also used in Firefox.

It's quite a bit of code, but much of it is relatively boring plumbing. The noteworthy parts are around the integration of the Rust code in the `c-dependencies/js-compute-runtime/rust-url/` directory, and the use if `cbindgen` for the FFI.