Make FetchEvent handling more spec-compliant (#8)
This introduces three significant changes:
1. Support multiple `"fetch"` event handlers instead of just one
2. Support `Promise<Response>` in addition to `Response` in `FetchEvent#respondWith`
3. Require `FetchEvent#respondWith` to be called synchronously from within a `"fetch"` handler

This last one is a breaking change: before, we supported invoking `respondWith` at any point. That isn't what the spec wants us to do, and it also means that we don't have a good way to determine when a `FetchEvent` has been fully handled. Our previous solution to that was to wait until all pending tasks have been completed, but that can potentially mean running an instance for much longer than the author intended.