Link without --allow-undefined and fix issues unconvered by doing so
Turns out `--allow-undefined` isn't necessary, and removing it means the linker doesn't emit imports from the `env` namespace for undefined symbols.

It also turns out that since the update to a Spidermonkey compiled with Rust 1.54.0, we weren't producing working binaries: the builds were importing a symbol (`mozalloc_abort`) from `env`, which is what happens with `--allow-undefined` if a symbol isn't found. I hadn't realized this problem because I did all my testing under Wizer, which is fine with unknown builtins as long as they aren't called.

This change will prevent bugs like this going forward, but we also clearly need a better testing story. I have a plan for that, but it's for another day.