Don't trap when trying to delete a non-existent header
As reported in #7, under certain circumstances we throw an internal error when calling `Headers#delete` with a non-existent header name. This only happens if the `Headers` object is associated with a `Request` or `Response`, and is caused by the hostcall we use to mirror the change to the host.

With this change, we only do the hostcall if the named header exists.

Fixes #7