Add support for including binary files, and for using typed arrays as `Response` bodies (#22)
It's often desirable to be able to embed binary assets into a wasm file. Eventually we'll have a standardized way to do so, but for now we don't.

This patch thus adds `fastly.includeBytes` as the moral equivalent to Rust's [include_bytes!](https://doc.rust-lang.org/1.54.0/std/macro.include_bytes.html) macro. `includeBytes` can only be used during the initialization/wizening phase, because that's the only time filesystem is (through the one-line change in `src/main.rs`) available.

The way binary data is embedded is pretty close to optimal: the data is without any intermediate copies placed into a buffer underlying the `Uint8Array` that `includeBytes` returns. That way, the data is part of the wizer snapshot without any overhead.

Making full use of these included assets requires the ability to also directly return them as responses, so this patch also adds support for passing typed array, `DataView`, or `ArrayBuffer` instances to the `Response` constructor, which in turn passes them on to the host without any intermediate copies.