Enable defining a default backend using `fastly.defaultBackend`
When using existing libraries, it can be difficult to pass in the backend to use in `fetch`. With this change, it's possible to define a default backend to use when none is passed in to the `fetch` call.

Note that this isn't ideal in situations where multiple backends are used. An alternative that's been discussed is to derive the backend from the request URL's host name instead. We might do that before the next SDK release.

In the meantime, this helps get WPT tests running.