Update to new version of spidermonkey-wasi-embedding
This pulls in a new version of the WASI embedding of SpiderMonkey, bringing new JSAPI surface for working with streams, and the ability to have release and debug builds built or downloaded side-by-side.

To work with the latter, this also includes a bunch of Makefile changes that make it easy to do debug builds, by defining the `DEBUG` env var.