Implement TransformStream
SpiderMonkey doesn't have a `TransformStream` implementation, and doing a proper one that'd work for Firefox as well would be a huge amount of work. So instead, we're adding our own implementation, using SpiderMonkey's `ReadableStream` and `WritableStream` implementations.

The goal is for the implementation to be safe, efficient, and spec-compliant, so we're not taking undue shortcuts here. The difference between this and an in-SpiderMonkey implementation is that we don't have to worry about multiple compartments and different kinds of cross-compartment wrappers, which simplifies the implementation significantly.

The implementation passes almost all spec tests (CI support for which will be added in a separate PR), and is I think reasonably efficient.