Support directly piping entire body streams to Request or Response objects instead of copying them by the chunk
This patch adds support for detecting when a Request or Response's body stream is piped to another Request or Response's body via a TransformStream. In that case, instead of reading every chunk from the body individually and then writing it to the destination body, we use the `xqd_body_append` hostcall to append the entire body in one go.

In situations where multiple bodies are piped to the same destination, this reduces the complexity from O(number of streams * number of chunks per stream) to O(number of streams), which in most cases basically takes JS execution out of the equation.

The patch also adds a whole bunch of type asserts, and includes some other small tweaks.