Improve handling of body streams in `{Request,Response}.text()` et al
The various methods on `Request` and `Response` that consume the `body` crashed when presented with content-provided `ReadableStreams`, and has some other suboptimal aspects. With this change, we still don't properly handle content-provided bodies in these methods, but at least we throw an exception when encountering them instead of crashing.

Most of this patch is refactoring things and laying the groundworks for actually handling content-provided streams. Apologies for the noisy patch.