Loose set of small fixes and improvements (#60)
* Update cargo dependencies and cargo-audit

* Run Promise reactions during initialization phase

Executing the top-level script might enqueue Promise reactions, which we should run before taking the snapshot. This is important both to ensure optimal performance—every action not taking during initialization needs to happen at runtime—but also for correctness: otherwise we'll only run the reactions *after* any listeners for the `fetch` event.

* Remove a redundant pair of error handling helpers

* Introduce abstraction for creating internal methods

For lots of spec algorithms, we need to create JS functions that can operate on an unforgeable receiver. This patch introduces an abstraction for creating these instead of doing it manually every time.

* Add a rudimentary method tracing facility