Cache more in CI (#112)
* Use rust-toolchain.toml in js-compute-runtime

I'm also only copying js-compute-runtime.wasm into the job's uploaded
artifacts, not any temporary files generated during the build.

* Cache Rust dependencies of js-compute-runtime

* Use rust-toolchain.toml at top-level

...and also rely on CI-provided versions of rustup, even on Mac.

* Share runner's $HOME with Centos container

The idea here is to use rustup and the Rust toolchain configured in the
outer container, while using the C toolchain from the inner Centos
container.

* Cache builds on `main` and pull requests

If `main` is uncached, then each new pull request starts with an empty
cache.

But ignore the build cache on tagged releases. This is possibly
excessive caution. It ensures that artifacts published with a release of
this SDK aren't corrupted by broken caches.