perf: switch to using a string_view for the base64 decoding and encoding functions to avoid unrequired copying of strings
This way we won't incur any cost for dropping data off of the front of the buffer like we did with `erase`, and we can ensure that none of the decoding functions will modify our buffer. We also avoid copying the input buffer when calling the decoding functions, as std::string_view is very cheap to copy around.

This excellent suggestion and explanation came from @elliottt - thank you for showing me string_view, it is the perfect type for this use-case.