move the building and publishing of the js-compute documentation site into this repository
previously this was in the compute-sdk-as-js repo and was a rust c@e application. I've moved the responsibility over into this repository and hooked it into the publishing of the @fastly/js-compute package so that they always stay in sync. I also changed from using a rust c@e app to a js c@e app for the site as that seems like a nice way for us to use our own software :-)