Automatically publish the npm package and associated documentation website
currently the publishing of both these things is done within the compute-sdk-as-js repository. as we are moving away from that repository and into this public one, I have added automation for publishing to npm when the @fastly/js-compute package has been updated by our release-please pull-requests.

The process is not yet fully automated for making a release as we still need to commit the built binaries into @fastly/js-compute - once we complete the work to avoid commiting the binaries then this can become fully automated!