Add integration/end-to-end testing into CI
This adds a series of jobs into our CI for deploying our test applications onto C@E and running their test suites.
Each application can defined a set-up and tear-down scripts, these are useful for setting up features such as Edge Dictionaries, ObjectStores, Logging Providers etc.

All the test applications have their own jobs which run in parallel and their individual test suites also run in parallel to help reduce the amount of time spent to the minimum possible.

The way that this works is by running a script which will:

    copy the fastly.toml.in file for each test applicat ion which can run it's tests on full C@E to fastly.toml
    delete any previously existing Fastly service for the application via fastly service delete --force --service-name $name_of_app
    build the wasm file, creating a new fastly service with a random domain, attaching the backends (if any) and publishing the wasm file to the new service via fastly compute publish -i (the backends are defined in the fastly.toml file in the [setup] section)
    run the test application's setup.js file if it exists
    wait until the random domain from step 3 resolves to the new fastly service
    run all the application's tests/assertions in parallel
    run the test application's teardown.js file if it exists
    delete the domain and service via fastly service delete --force
    exit with 1 if any step above or assertion failed or exit with 0 if everything passed

Where appropriate, a step will be retried automatically up to 10 times, this is useful to avoid any temporary issues which can sometimes occur when doing end-to-end testing and deployments.

The script can also be used on a developer's machine which can be useful for the situations where viceroy tests are not possible (such as right now for ObjectStore tests) by running ./integration-tests/js-compute/test.js <fixture-names...>