only enforce licenses for runtime dependencies and not development dependencies
previously we had have to relax the allow-licenses field due to a development dependency having a license which was not in the allow-list.

The newer version of dependency-review-action now has the option to specify which 'scope' to check the licenses for. I've set the scope to be 'runtime', which means it will no longer check the development dependency licenses.

In a future commit we can revert the allow-licenses field back to it's original, stricter list.