Update to version 0.5.2 of js-compute-runtime
This version of js-compute-runtime includes  a security fix for the `crypto.getRandomValues` method to ensure it always uses sufficiently random values. The previous version when compiled to WebAssembly, would use the same seed value for each execution of the WebAssembly module, meaning the sequence of numbers generated was predictable for a specific WebAssembly module. The new version uses a different seed value for each execution.