Wrap wasm-opt to prevent clang from running while linking (#350)
Clang now runs wasm-opt automatically if it's present in the path when linking wasm artifacts. For debug builds on osx this produces a bus error, it's also not clear that we want wasm-opt -O2 applied to our debug artifacts.

This PR introduces a script named wasm-opt that immediately exits successfully, and adds it to the PATH when linking the runtime. This prevents clang from running the real wasm-opt, and gives us more control over what optimizations are enabled when we do choose to run it.