feat: implement SubtleCrypto.prototype.digest method
This commit is the start towards full SubtleCrypto support in the runtime.

I've implemented the `SubtleCrypto` class but with only one method so far, `SubtleCrypto.prototype.digest`.

I've made `globalThis.crypto` become an instance of `Crypto` and have added an instance of `SubtleCrypto` as a getter on `Crypto.prototype.subtle` as per the [WebIDL specification](https://w3c.github.io/webcrypto/#webidl-1120328020):
```webidl
partial interface mixin WindowOrWorkerGlobalScope {
  [SameObject] readonly attribute Crypto crypto;
};

[Exposed]
interface Crypto {
  [SecureContext] readonly attribute SubtleCrypto subtle;
  ArrayBufferView getRandomValues(ArrayBufferView array);
  [SecureContext] DOMString randomUUID();
};
```

Currently 76 out of 80 of the Web Platform Tests pass for `SubtleCrypto.prototype.digest`, the ones which fail are to do with providing a large `data` parameter and altering the `data` parameter's buffer after calling `SubtleCrypto.prototype.digest`. I'm not sure yet as to why these tests are failing, it is confusing for me as the same test but with a smaller `data` parameter does pass as expected.