fix: ensure CacheOverride bitflags are the same value as definde in xqd
Previously we had defined a `http-cache-override-tag` flag of `none` - which is not a defined flag within xqd. This extra defined flag meant that all our flag values were off-by-one.

This commit removes the `none` flag and also adds static_asserts to ensure that our world-adapter values are the same values that we have defined in cache-override.h

This commit also remove the convert_tag function which was used in the definition of `xqd_fastly_http_req_cache_override_set` when in non-component-mode to convert the provided tag into a single flag value. That behaviour is incorrect as it effectively changes the behaviour of the flag into an enum, whereas flags are different to enums because multiple flags can be set at the same time.