chore: allow multiple PRs to run the end-to-end tests at the same time
This commit updates the test script to include the branch name in each of the end-to-end test applications -- this allows multiple branches to run the end-to-end tests at the same time.

I've also updated the script to not tear-down the service if the tests have failed, this will enable us to debug the deployed application. When the tests pass, the tear-down will happen.