feat: implement Fanout for JS SDK
Below is a product description of Fanout:

------

Fanout is a publish/subscribe message broker operating at the Fastly edge, powered by [Pushpin](https://pushpin.org/). This makes it easy to build and scale realtime/streaming applications which push data instantly to browsers, mobile apps, servers, and other devices.

Without Fanout, clients make short-lived HTTP requests to Fastly, and Fastly responds with a resource from cache or an origin server. When a request is passed to Fanout, it keeps that connection open indefinitely, assigns the connection to one or more channels, and waits for data to be published to those channels. Origin servers or Fastly services may use an API to publish data to a channel, and Fanout will distribute that data to all the clients which have a subscription to that channel.

Any regular HTTP request may be upgraded into an event-driven response using a transport such as Server-Sent-Events, or Long-Polling. If the request from the client is a WebSocket request, the resulting stream is bidirectional.

This architecture has some nice advantages over more proprietary streaming services:

- Integrated into your Fastly service, so any pre-processing you do to requests to your Fastly service (e.g. for authentication), you can do to streaming requests too.
- Any HTTP client or server can be used - including serverless / functions-as-a-service.
- Any HTTP response can be turned into an evented stream (e.g. a progressive JPEG, a log file, or an API endpoint that performs asynchronous operations).
- No need to use a separate domain for streaming data.

-----

Most JavaScript Compute Services have an entry function which looks like this:
```js
addEventListener("fetch", event => event.respondWith(handleRequest(event)));
```
Where `handleRequest` contains the logic for the application.

With this in mind, I wanted to design a Fanout API which works with this type of entry function.
What I landed on was a function which returns a JS Response, this JS Response has some internal slots set to indicate it's a Fanout response, if the JS Response is then given to `event.respondWith` then instead of calling `fastly_http_resp_send_downstream` we call `fastly_http_req_redirect_to_grip_proxy` which is the Fanout hostcall.

This is what it looks like in an application:
```js
import { createFanoutHandoff } from "fastly:fanout";

addEventListener("fetch", (event) => event.respondWith(handleRequest(event)));

async function handleRequest(event) {
  try {
    const url = new URL(event.request.url);
    if (url.pathname === '/stream') {
      return createFanoutHandoff(event.request, 'fanout');
    } else {
      return new Response('oopsie, make a request to /stream for some fanout goodies', { status: 404 });
    }
  } catch (error) {
    console.error({error});
    return new Response(error.message, {status:500})
  }
}
```

Co-authored-by: Trevor Elliott <telliott@fastly.com>