Add option to use `weval`, an experimental JS AOT option using partial evaluation.
This PR pulls in some still-experimental work to partially evaluate the
SpiderMonkey JS interpreter together with JS scripts, ahead-of-time, to
produce compiled code.

It should not yet be considered production-worthy, pending more testing
and proving out its stability and performance; but, it is at a point
where we believe it should be under a *non-default* option so that users
of the SDK may try it. No guarantees! It may explode your bytecode!

The support works by bundling two builds of the engine Wasm -- with and
without the weval intrinsics -- and invoking either Wizer with the
"normal" build or weval with the "weval" build, as appropriate. The
`--enable-weval` option selects the latter.