Abstract over async handles (#536)
Introduce an `AsyncHandle` type to the host api, and refactor the event loop to use them instead.

I reworked the interface to the underlying `fastly_async_io_select` host call, so that it's now a bit more clear when the timeout has expired. I've also added comments to `process_pending_async_tasks` to better explain the interactions between the two action queues.