chore: Remove more direct uses of `free`, replacing them with unique ptrs (#542)
Remove all but three direct uses of `JS_free`, replacing them with either `mozilla::UniquePtr`, or `std::vector`. The resulting pattern is that direct calls to `JS_free` are no longer needed, and functions that take ownership of memory now need to use calls to `.release()` on the unique pointers.

I also noticed that I missed a call-site to `fastly_http_req_uri_get` when converting over to the host_api, so I've updated that as well.