fix: only apply our pipe_to/pipe_through optimisations to TransformStreams who have no transformers.
We have a performance optimisation for when the receiver is a native source, whereby we append the entire source body to the destination using a single native hostcall, this greatly improves the performance (https://barstool.engineering/a-real-world-comparison-between-cloudflare-workers-and-fastly-compute-edge/)

This optimisation however, should only be applied when the TransformStream is being used in this particular, having not supplied a transformer to the TransformStream constuctor. If we apply the optimisation when a transformer is supplied, the transformer would never get called, which is not correct behavior.