Replace RAII with a macro
We can't make use of descructors to ensure the transaction is cancelled because the JS API we expose uses an asynchronous callback and we have to keep the transaction open until the callback resolves, which can happen over many JS event-loop ticks. If we used RAII, the object would be destructed before the callback is called (because the callback will be called on a separate tick) and the handle would have become invalidated too early.

We could use RAII within `getOrSetThenHandler` but not within the 'parent' `getOrSet`. I opted to use the same logic (macro) for both functions, rather than use RAII for one and a macro for the other.