refactor: consolidate all test applications into one test application
by merging the test applications into one, we are able to run all the test suites locally much faster than before because it involves compiling to wasm once and spinning up viceroy once, compared to the ~50 wasms and viceroys we were using before.

It should also help reduce the amount of flakes we see when publishing to Fastly and reduce some constraints on internal Fastly systems.