AOT compilation: rename flag; no longer experimental.
The AOT compilation option is now no longer advertised as experimental;
it is ready for general use for anyone who wishes to try it. It is still
off by default, i.e., an opt-in feature.

This PR renames the flag to `--enable-aot`, and continues to accept the
old flag spelling (`--enable-experimental-aot`) while printing a warning
that it has been renamed.