Reject core cache body ranges where the end precedes the start
These are rejected in Go as of
https://github.com/fastly/compute-sdk-go/pull/165,
and in Rust as of PR 5097 in that repository.

This is a guest (SDK/runtime) change rather than a host change so as not
to break users that may accidentallly rely on this "feature"; hopefully
their next SDK upgrade will highlight the issue.