Option to respect range during cache streaming
The core cache API allows for cache items to be concurrently streamed to / from the cache.
If:
- The size of the cached item's body was not provided by the writer
- The reader requests a specific range of the cached item's body (`to_stream_from_range`)
- The writer and reader are concurrent, i.e. the body is streamed from one to the other
the core cache API will ignore the requested range and provide the whole body.
There is no explicit notification that the whole body is provided instead of a range.

In this SDK release, this remains the default behavior. However, lookup calls (lookup, transaction lookup) now offer an `always_use_requested_range` option. If set to true, body reads conducted as part of this lookup/transaction/replacement will always use the requested range, even when streaming.

In a future (major) SDK release, the default behavior will change to `always_use_requested_range`.